# (#3921) - avoid extend() in merge.js