Same as #4140, but containing just the (risky) in-place mangling of trees/paths.

This is a harmless commit per our unit tests, but since it introduces slightly surprising behavior (`merge()` modifies the input objects), it needs a lot of justification from profiling. When I get a chance (or if someone else wants to hop on this!) I think I can demonstrate that this offers a big boost to our speed (especially in `bulkDocs()`), so it's well worth it.