@nolanlawson 2015-08-14
OK, I added a new performance test, because none of our current ones were capturing what I wanted. This new one is called `basic-updates`, and it just inserts 100 documents and then updates them each 100 times.

This proves that this PR is an enormous performance boost. I was actually really surprised by how big it turned out to be. All tests are on my crappy Acer Aspire Linux laptop.

| Environment | Before | After |
| ---- | ---- | --- |
| iojs 2.5 (leveldb) | 40365ms | 15350ms |
| Chrome 44 (idb) | 70521ms | 40288ms |
| Firefox 41 (idb) | 65415ms | 34425ms |
| Chrome 11 (websql) | 57045ms | 41881ms |

I would like to test in Safari to see how JSCore is affected, but I don't have an Apple device handy.