@daleharvey 2015-08-14
Heh, ok with that type of speedup then big :+1: 

Same comment about the test failures as in the other PR