@nolanlawson 2015-08-15
Yeah, the idb-alt and localstorage-down tests are permafail; must be some module that they both depend on. Gonna wait it out and just cross my fingers for now.